+++ src/dbg/debugger_commands.h
@@ -58,6 +58,17 @@ CMDRESULT cbDebugSetBPXMemoryFastResume(int argc, char* argv[]);
 CMDRESULT cbDebugSetBPXMemorySingleshoot(int argc, char* argv[]);
 CMDRESULT cbDebugSetBPXMemorySilent(int argc, char* argv[]);
 CMDRESULT cbDebugResetBPXMemoryHitCount(int argc, char* argv[]);
+CMDRESULT cbDebugSetBPXDLLName(int argc, char* argv[]);
+CMDRESULT cbDebugSetBPXDLLCondition(int argc, char* argv[]);
+CMDRESULT cbDebugSetBPXDLLLog(int argc, char* argv[]);
+CMDRESULT cbDebugSetBPXDLLLogCondition(int argc, char* argv[]);
+CMDRESULT cbDebugSetBPXDLLCommand(int argc, char* argv[]);
+CMDRESULT cbDebugSetBPXDLLCommandCondition(int argc, char* argv[]);
+CMDRESULT cbDebugGetBPXDLLHitCount(int argc, char* argv[]);
+CMDRESULT cbDebugSetBPXDLLFastResume(int argc, char* argv[]);
+CMDRESULT cbDebugSetBPXDLLSingleshoot(int argc, char* argv[]);
+CMDRESULT cbDebugSetBPXDLLSilent(int argc, char* argv[]);
+CMDRESULT cbDebugResetBPXDLLHitCount(int argc, char* argv[]);
 CMDRESULT cbDebugBplist(int argc, char* argv[]);
 CMDRESULT cbDebugStepInto(int argc, char* argv[]);
 CMDRESULT cbDebugeStepInto(int argc, char* argv[]);
