+++ src/dbg/x64_dbg_dbg.vcxproj.filters
@@ -350,6 +350,9 @@
     <ClCompile Include="animate.cpp">
       <Filter>Source Files\Utilities</Filter>
     </ClCompile>
+    <ClCompile Include="breakpoint_commands.cpp">
+      <Filter>Source Files\Debugger Core</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="x64_dbg.h">
